Fix problem with loading commits with unchanged tracked objects (#2)
When resetting to a commit, the version manager will look through for
the most recent commit with a commited change for each object when
rebuilding the scene. This is to prevent the case where objects'
positions might not be able to be resolved if they weren't changed in
the very commit that is being reset to.