Integrate Commit design with VersionManager
Commit objects are essentially link list nodes (I called them ChainLink)
with some other commit data such as the message and the versioning data
required to reconstructed the commit.

The VersionControllers each hold a running "version" instead of using
the commit ids in the History objects. This is so that each object can
have it's own independent contiguous version histories. The Commit
objects contains the pairing between VersionController and version to
rebuild a given commit.