Implement Commit objects; integrate with VersionManager, tests (#4)
Commit objects are essentially linked list nodes (I called them ChainLink)
with some other commit data such as the message and the versioning data
required to reconstruct the commit.

The VersionControllers each hold a running "version" instead of using
the commit IDs in the History objects. This is so that each object can
have it's own independent contiguous version histories. The Commit
object contains the pairing between VersionController and version to
rebuild a given commit.